+++ src/index.ts
@@ -1,3 +1,2 @@
 export { ScrollbarModule } from './scrollbar/scrollbar.module';
 export { ScrollbarComponent } from './scrollbar/scrollbar.component';
-export { ScrollBarEvent } from './scrollbar/scrollbar.model'